// frontend/lib/api/model/setCreate.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Workout API
 * A workout tracking API built with FastAPI
 * OpenAPI spec version: 1.0.0
 */
import type { SetCreateNoteText } from "./setCreateNoteText";

/**
 * Schema for creating a new set.
 */
export interface SetCreate {
  /**
   * Number of clean repetitions completed
   * @minimum 0
   */
  clean_reps: number;
  /**
   * Number of forced repetitions completed
   * @minimum 0
   */
  forced_reps?: number;
  /** Optional notes about this set */
  note_text?: SetCreateNoteText;
  /**
   * Weight used for this set (in pounds)
   * @minimum 0
   */
  weight: number;
}